Use valid url value in library.properties
The previous library.properties url field value perhaps anticipates an eventual reference page for the library, but at
present there is no such page, so it provides a poor experience for the user. Better to point them to a site that exists
now. Then update at such time as a reference is created.